Крупноапдейт.
Выкинуты ненужные флаги, доделаны нужные флаги, пофикшена кровь на предметах, фикс С4, ещё фиксы и ещё фиксы.

Допилена аномалистика методом выпила её к хуям и впила актуальной с Бея. Допилил уже Бейскую, теперь аномашинки можно пересобрать с деталями из РнД для повышения эффективности. Спрайты мои, критика принимается только в комплекте со спрайтами получше.

На астероиде теперь можно выкопать артефакт или ископаемое.

Добавлена тёплая ламповая зарядка стволов а-ля Флег. Русскую рулетку теперь можно с любым револьвером, даже с детективским, хотя смертельность резиновой пули в висок не гарантируется.

+++ code/game/objects/effects/effect_system.dm
@@ -11,14 +11,14 @@ would spawn and follow the beaker, even if it is carried or thrown.
 	icon = 'icons/effects/effects.dmi'
 	mouse_opacity = 0
 	unacidable = 1//So effect are not targeted by alien acid.
-	flags = TABLEPASS
+	flags = 0
 
 /obj/effect/effect/water
 	name = "water"
 	icon = 'icons/effects/effects.dmi'
 	icon_state = "extinguish"
 	var/life = 15.0
-	flags = TABLEPASS
+	pass_flags = PASSTABLE
 	mouse_opacity = 0
 
 /obj/effect/effect/smoke
@@ -325,8 +325,9 @@ steam.start() -- spawns the effect
 					sleep(10)
 					step(smoke,direction)
 				spawn(75+rand(10,30))
-					smoke.delete()
-					src.total_smoke--
+					if(smoke)
+						smoke.delete()
+						src.total_smoke--
 
 
 /////////////////////////////////////////////
